refactor: make use of Babel (#488)
* refactor: rename lib to src

* refactor: use ESM

* chore(package.json): add prepublish task

* chore: rename lib to dist
