--- lib/DevMiddlewareError.js
@@ -1,5 +1,0 @@
-'use strict';
-
-class DevMiddlewareError extends Error {}
-
-module.exports = DevMiddlewareError;
