--- lib/utils/index.js
@@ -1,23 +1,0 @@
-'use strict';
-
-const getFilenameFromUrl = require('./getFilenameFromUrl');
-const handleRangeHeaders = require('./handleRangeHeaders');
-const ready = require('./ready');
-const reporter = require('./reporter');
-const setupHooks = require('./setupHooks');
-const setupLogger = require('./setupLogger');
-const setupOutputFileSystem = require('./setupOutputFileSystem');
-const setupRebuild = require('./setupRebuild');
-const setupWriteToDisk = require('./setupWriteToDisk');
-
-module.exports = {
-  getFilenameFromUrl,
-  handleRangeHeaders,
-  ready,
-  reporter,
-  setupHooks,
-  setupLogger,
-  setupOutputFileSystem,
-  setupRebuild,
-  setupWriteToDisk,
-};
