+++ package.json
@@ -7,26 +7,28 @@
   "author": "Tobias Koppers @sokra",
   "homepage": "https://github.com/webpack/webpack-dev-middleware",
   "bugs": "https://github.com/webpack/webpack-dev-middleware/issues",
-  "main": "lib/index.js",
+  "main": "dist/index.js",
   "engines": {
     "node": ">= 10.13"
   },
   "scripts": {
     "commitlint": "commitlint --from=master",
     "lint:prettier": "prettier \"{**/*,*}.{js,json,md,yml,css}\" --list-different",
-    "lint:js": "eslint --cache lib test",
+    "lint:js": "eslint --cache src test",
     "lint": "npm-run-all -l -p \"lint:**\"",
+    "build": "del dist && babel src -d dist",
     "release": "standard-version",
     "security": "npm audit",
     "test:only": "cross-env NODE_ENV=test jest",
     "test:watch": "cross-env NODE_ENV=test jest --watch",
-    "test:coverage": "cross-env NODE_ENV=test jest --collectCoverageFrom=\"lib/**/*.js\" --coverage",
+    "test:coverage": "cross-env NODE_ENV=test jest --collectCoverageFrom=\"src/**/*.js\" --coverage",
     "pretest": "npm run lint",
     "test": "cross-env NODE_ENV=test npm run test:coverage",
-    "defaults": "webpack-defaults"
+    "defaults": "webpack-defaults",
+    "prepublish": "npm run build"
   },
   "files": [
-    "lib"
+    "dist"
   ],
   "peerDependencies": {
     "webpack": "^4.0.0"
