+++ src/DevMiddlewareError.js
@@ -1,0 +1,3 @@
+class DevMiddlewareError extends Error {}
+
+export default DevMiddlewareError;
