+++ src/utils/ready.js
@@ -1,7 +1,5 @@
-'use strict';
-
 // eslint-disable-next-line consistent-return
-function ready(context, fn, req) {
+export default function ready(context, fn, req) {
   if (context.state) {
     return fn(context.webpackStats);
   }
@@ -9,5 +7,3 @@ function ready(context, fn, req) {
   context.log.info(`wait until bundle finished: ${req.url || fn.name}`);
   context.callbacks.push(fn);
 }
-
-module.exports = ready;
